Extract base Cooconed class and plugin definition to ES5 classes and mixins
