+++ npm/src/plugins.js
@@ -1,0 +1,7 @@
+import limitMixin from "./plugins/limit"
+import reorderableMixin from "./plugins/reorderable"
+
+export {
+  limitMixin,
+  reorderableMixin
+}
